Fix MaximumMatching algorithm and add tests
Fixed the MaximumMatching algorithm. Added tests for MaximumMatching and
PerfectMatching.

Currently, MaximumMatching tests are being generated by
fixing the number of odd components.

+++ code/algorithms/rank2/rank2-algorithm.hpp
@@ -31,7 +31,7 @@ class Rank2AlgorithmStrategy : public IAlgorithmStrategy {
    * Complexity: O(n^4).
    */
   vector<pair<int, int>> solve(const Graph& G) const override {
-    const size_t n = V(G).size(), m = E(G).size();
+    const int n = V(G).size(), m = E(G).size();
 
     auto T = GetTutteMatrix(G);
     if (T.isSingular()) {
